Flytter ut informasjon om samtykke i egen fil
